backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
openai>=1.40.0
httpx>=0.28.1
requests==2.32.3
pandas==2.2.0
pydantic==2.6.1
python-multipart==0.0.9

# Celonis SDK (Needs extra index url during pip install, e.g. pip install -r requirements.txt --extra-index-url=https://pypi.celonis.cloud/)
pycelonis
